Extract MarkerUtils
